wednesday-api/requirements.txt
Flask==3.0.3
Mastodon.py==1.8.1
